Change for addition of tambourine sound
The original sixteen sounds included a finger snap.  It was supposed to be a tambourine.  Changes have to be made to both the sound card that goes into the Robertson Wav Trigger and the sound control Arduino Nano.  The sound card gets the 16 bit PCM 41400 sample rate file and the control Nano gets the name for its list of drum sounds.  The tambourine is sound 13, so the sound file name is 013_Tambourine.wav.  If the name is not changed in the control file everything will work but you will need to remember what each drum name actually represents when you select different drums.